requirements.txt
bson
flask
pymongo
python-Levenshtein